create SAVE(controller, repository, dto, service) logic of only boardTest and test dummy code

+++ src/main/java/com/devjck/springboard/dto/board/BoardResponseDto.java
@@ -5,7 +5,7 @@ import lombok.Getter;
 
 @Getter
 public class BoardResponseDto {
-    private int boardSeq;
+    private Long boardSeq;
 
     private int boardWriterId;
